Fix design flaw in Monte Carlo Tree Search
Flaw was the way it was scoring the games in the tree.
Was always from the perspective of the player on the root node.
This is bad because it affects the selection step.
We need to update the score relative to which player just moved.
That way, when calculating the UCT and selecting a node, we
assume that each player is making the best move relative to
themself. Previously we were essentially assuming that the
opponent was going to make the best move for us!